mips: Fix usage of configTICK_RATE_HZ
TICK_RATE_HZ is the frequency of the timer interrupt (eg 1KHz or 100Hz) not the
frequency of the timer/counter used to generate the interrupt.

We have added a new config option 'configTIMER_RATE_HZ' for the freqeunce of the
MIPS count/compare timer (which is typically core clock / 2).